source/actuator_service/requirements.txt
python-qpid-proton==0.39.0
requests==2.31.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
sse-starlette==2.1.0
httpx
redis